+++ light-client/src/light_client.rs
@@ -13,7 +13,7 @@ use crate::{
     bail,
     errors::{Error, ErrorKind},
     state::State,
-    types::{Height, LightBlock, PeerId, Status, Time, TrustThreshold},
+    types::{Height, LightBlock, PeerId, Status, TrustThreshold},
 };
 
 /// Verification parameters
@@ -26,23 +26,16 @@ pub struct Options {
     /// and trusted validator set is sufficient for a commit to be
     /// accepted going forward.
     pub trust_threshold: TrustThreshold,
+
     /// How long a validator set is trusted for (must be shorter than the chain's
     /// unbonding period)
     pub trusting_period: Duration,
+
     /// Correction parameter dealing with only approximately synchronized clocks.
     /// The local clock should always be ahead of timestamps from the blockchain; this
     /// is the maximum amount that the local clock may drift behind a timestamp from the
     /// blockchain.
     pub clock_drift: Duration,
-    /// The current time
-    pub now: Time,
-}
-
-impl Options {
-    /// Override the stored current time with the given one.
-    pub fn with_now(self, now: Time) -> Self {
-        Self { now, ..self }
-    }
 }
 
 /// The light client implements a read operation of a header from the blockchain,
@@ -154,22 +147,20 @@ impl LightClient {
         state: &mut State,
     ) -> Result<LightBlock, Error> {
         // Let's first look in the store to see whether we have already successfully verified this block
-        if let Some(light_block) = state.light_store.get(target_height, Status::Verified) {
+        if let Some(light_block) = state.light_store.get_trusted_or_verified(target_height) {
             return Ok(light_block);
         }
-
-        // Override the `now` fields in the given verification options with the current time,
-        // as per the given `clock`.
-        let options = self.options.with_now(self.clock.now());
 
         let mut current_height = target_height;
 
         loop {
+            let now = self.clock.now();
+
             // Get the latest trusted state
             let trusted_state = state
                 .light_store
                 .latest_trusted_or_verified()
-                .ok_or_else(|| ErrorKind::NoInitialTrustedState(Status::Verified))?;
+                .ok_or_else(|| ErrorKind::NoInitialTrustedState)?;
 
             if target_height < trusted_state.height() {
                 bail!(ErrorKind::TargetLowerThanTrustedState {
@@ -179,11 +170,10 @@ impl LightClient {
             }
 
             // Check invariant [LCV-INV-TP.1]
-            if !is_within_trust_period(&trusted_state, options.trusting_period, options.now) {
+            if !is_within_trust_period(&trusted_state, self.options.trusting_period, now) {
                 bail!(ErrorKind::TrustedStateOutsideTrustingPeriod {
                     trusted_state: Box::new(trusted_state),
-                    options,
-                    status: Status::Verified
+                    options: self.options,
                 });
             }
 
@@ -195,28 +185,31 @@ impl LightClient {
                 return Ok(trusted_state);
             }
 
-            // Fetch the block at the current height from our peer
-            let current_block = self.get_or_fetch_block(current_height, state)?;
+            // Fetch the block at the current height from the light store if already present,
+            // or from the primary peer otherwise.
+            let (current_block, status) = self.get_or_fetch_block(current_height, state)?;
 
             // Validate and verify the current block
             let verdict = self
                 .verifier
-                .verify(&current_block, &trusted_state, &options);
+                .verify(&current_block, &trusted_state, &self.options, now);
 
             match verdict {
                 Verdict::Success => {
-                    // Verification succeeded, add the block to the light store with `verified` status
-                    state.light_store.update(&current_block, Status::Verified);
+                    // Verification succeeded, add the block to the light store with
+                    // the `Verified` status or higher if already trusted.
+                    let new_status = Status::most_trusted(Status::Verified, status);
+                    state.light_store.update(&current_block, new_status);
                 }
                 Verdict::Invalid(e) => {
-                    // Verification failed, add the block to the light store with `failed` status, and abort.
+                    // Verification failed, add the block to the light store with `Failed` status, and abort.
                     state.light_store.update(&current_block, Status::Failed);
 
                     bail!(ErrorKind::InvalidLightBlock(e))
                 }
                 Verdict::NotEnoughTrust(_) => {
                     // The current block cannot be trusted because of missing overlap in the validator sets.
-                    // Add the block to the light store with `unverified` status.
+                    // Add the block to the light store with `Unverified` status.
                     // This will engage bisection in an attempt to raise the height of the highest
                     // trusted state until there is enough overlap.
                     state.light_store.update(&current_block, Status::Unverified);
@@ -230,35 +223,33 @@ impl LightClient {
         }
     }
 
-    /// Look in the light store for a block from the given peer at the given height.
-    /// If one cannot be found, fetch the block from the given peer.
+    /// Look in the light store for a block from the given peer at the given height,
+    /// which has not previously failed verification (ie. its status is not `Failed`).
+    ///
+    /// If one cannot be found, fetch the block from the given peer and store
+    /// it in the light store with `Unverified` status.
     ///
     /// ## Postcondition
     /// - The provider of block that is returned matches the given peer.
-    #[post(ret.as_ref().map(|lb| lb.provider == self.peer).unwrap_or(true))]
+    #[post(ret.as_ref().map(|(lb, _)| lb.provider == self.peer).unwrap_or(true))]
     pub fn get_or_fetch_block(
         &self,
-        current_height: Height,
+        height: Height,
         state: &mut State,
-    ) -> Result<LightBlock, Error> {
-        let current_block = state
-            .light_store
-            .get(current_height, Status::Verified)
-            .or_else(|| state.light_store.get(current_height, Status::Unverified));
+    ) -> Result<(LightBlock, Status), Error> {
+        let block = state.light_store.get_non_failed(height);
 
-        if let Some(current_block) = current_block {
-            return Ok(current_block);
+        if let Some(block) = block {
+            return Ok(block);
         }
 
-        self.io
-            .fetch_light_block(self.peer, AtHeight::At(current_height))
-            .map(|current_block| {
-                state
-                    .light_store
-                    .insert(current_block.clone(), Status::Unverified);
+        let block = self
+            .io
+            .fetch_light_block(self.peer, AtHeight::At(height))
+            .map_err(ErrorKind::Io)?;
 
-                current_block
-            })
-            .map_err(|e| ErrorKind::Io(e).into())
+        state.light_store.insert(block.clone(), Status::Unverified);
+
+        Ok((block, Status::Unverified))
     }
 }